feat(bubbles-ui): URL 駆動の世界線遷移を全バブリで使えるように
bublys-os だけが持っていた「ブラウザ URL ↔ root universe apex」のバインドを、
任意の SnapshotCodec を取る lib hook として昇格し、各バブリ app が同等の
view 遷移機能を 1 行で得られるようにする。

lib（bubbles-ui）:
- 新 hook: world-line/useBrowserRootArrangementWorldLine(codec)
  OS 側の useRootArrangementWorldLine と同内容（apex → pushState、popstate →
  moveTo、self-heal、訪問トレイル）。`<base>` の指定だけ呼び出し側から注入
- BublyStoreProvider に 3 つの optional props を追加:
  - enableWorldLine: initWorldLineGraph() を 1 回だけ呼ぶ
  - domainRegistry:  DomainRegistryProvider で内側を包む
  - urlBinding:      useBrowserRootArrangementWorldLine を起動
  これらを揃えて指定するとバブリは bublys-os と同等の URL 遷移を獲得
- 内側組み立て順: PersistGate > DomainRegistryProvider > UrlBindingHost > children

bublys-os:
- useRootArrangementWorldLine は薄いシム（rootBrowserSnapshotCodec を渡すだけ）
  に縮小。ロジックは lib 側に一本化

各バブリの standalone app（gakkai-shift / ekikyo / sekaisen-igo / tailor-genie /
shift-puzzle）の BublyStoreProvider に enableWorldLine / domainRegistry /
urlBinding={makeSnapshotCodec("universe")} を追加。手書きの initWorldLineGraph()
呼び出しは Provider に集約されたので削除。これでバブリのスタンドアロン URL も
`/universe@<node>` 形式になり、ブラウザ戻る/進む & ディープリンクが効く。

dynamic bubly（OS にスクリプト注入される形）は BublyStoreProvider を経由しない
ので、OS のバインドと競合しない（自然に成立）。
