+++ bublys-libs/bubbles-ui/src/lib/world-line/index.ts
@@ -1,2 +1,3 @@
 export * from "./bubbleArrangementDomain.js";
 export * from "./useUniverseArrangementWorldLine.js";
+export * from "./useBrowserRootArrangementWorldLine.js";
