+++ shift-puzzle-bubly/shift-puzzle-app/src/app/app.tsx
@@ -1,7 +1,13 @@
 import PeopleIcon from '@mui/icons-material/People';
 import TaskIcon from '@mui/icons-material/Task';
 import GridOnIcon from '@mui/icons-material/GridOn';
-import { BublyApp, BublyStoreProvider, BubbleRouteRegistry } from '@bublys-org/bubbles-ui';
+import {
+  BublyApp,
+  BublyStoreProvider,
+  BubbleRouteRegistry,
+  BUBBLE_ARRANGEMENT_DOMAIN,
+  makeSnapshotCodec,
+} from '@bublys-org/bubbles-ui';
 
 // shift-puzzle-libs のslicesをimport（自動注入される）
 import '@bublys-org/shift-puzzle-libs';
@@ -23,6 +29,9 @@ export function App() {
     <BublyStoreProvider
       persistKey="shift-puzzle-standalone"
       initialBubbleUrls={['shift-puzzle/shift-plans', 'shift-puzzle/tasks']}
+      enableWorldLine
+      domainRegistry={BUBBLE_ARRANGEMENT_DOMAIN}
+      urlBinding={makeSnapshotCodec('universe')}
     >
       <BublyApp
         title="シフトパズル"
